requirements.txt
streamlit
pandas
streamlit-lightweight-charts
google-generativeai 